# OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "WebApiAutores",
    "description": "Api de Autores, Libros y Comentarios V1",
    "contact": {
      "name": "Sergio Britos",
      "url": "https://google.com",
      "email": "trojan.v6@gmail.com"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/V1/SOAP_gRPC_OpenAPI/soap": {
      "get": {
        "tags": [
          "SOAP_gRPC_OpenAPI"
        ],
        "operationId": "SOAP_Service",
        "parameters": [
          {
            "name": "incluirHATEOAS",
            "in": "header",
            "style": "simple",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AutorDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AutorDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AutorDTO"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AutorDTO"
                  }
                }
              },
              "text/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AutorDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/V1/SOAP_gRPC_OpenAPI/grpc": {
      "get": {
        "tags": [
          "SOAP_gRPC_OpenAPI"
        ],
        "operationId": "Grpc_Service",
        "parameters": [
          {
            "name": "incluirHATEOAS",
            "in": "header",
            "style": "simple",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AutorDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AutorDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AutorDTO"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AutorDTO"
                  }
                }
              },
              "text/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AutorDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/V1/SOAP_gRPC_OpenAPI/openAPI": {
      "get": {
        "tags": [
          "SOAP_gRPC_OpenAPI"
        ],
        "operationId": "OpenAPI_Service",
        "parameters": [
          {
            "name": "incluirHATEOAS",
            "in": "header",
            "style": "simple",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AutorDTO": {
        "type": "object",
        "properties": {
          "enlaces": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DatoHATEOAS"
            },
            "nullable": true
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "nombreCompleto": {
            "type": "string",
            "nullable": true
          },
          "imagen": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DatoHATEOAS": {
        "type": "object",
        "properties": {
          "enlace": {
            "type": "string",
            "nullable": true
          },
          "descripcion": {
            "type": "string",
            "nullable": true
          },
          "metodo": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}